+++ cmd/tmpl/tmpl.go
@@ -1,0 +1,5 @@
+package tmpl
+
+import "html/template"
+
+var Tpl *template.Template
