+++ cmd/version/version.go
@@ -1,0 +1,3 @@
+package version
+
+const Version string = "version 0.1.1"
