+++ deployments/docker-compose/docker-compose.yaml
@@ -1,6 +1,6 @@
-version: '2.1'
+version: '3.1'
 services:
   go-tooling-portal:
-    image: docker.io/sfarosu/go-tooling-portal:latest
+    image: sfarosu/go-tooling-portal:latest
     ports:
       - "8080:8080"
