+++ go.mod
@@ -1,5 +1,5 @@
 module github.com/sfarosu/go-tooling-portal
 
-go 1.13
+go 1.16
 
-require github.com/prometheus/client_golang v1.3.0
+require github.com/prometheus/client_golang v1.11.0
