+++ web/templates/index.html
@@ -45,6 +45,9 @@
           <li class="nav-item">
             <a class="nav-link" href="/ssh">SSHKeyGenerator</a>
           </li>
+          <li class="nav-item">
+            <a class="nav-link" href="/jsonprettify">JsonPrettifier</a>
+          </li>
         </ul>
       </div>
       </nav>
@@ -68,8 +71,8 @@
         <div class="row">
           <div class="col-md-6">
             <h2 class="text-primary">HTPasswordGenerator</h2>
-            <p>Use htpasswd to encrypt passwords. Choose between 3 algorithms : </p>
-            <h6>-apr1 : MD5-based password algorithm, Apache variant (for Openshift)</h6>
+            <p>Use htpasswd to encrypt passwords. Choose between 3 algorithms: </p>
+            <h6>-apr1 : MD5-based password algorithm, Apache variant</h6>
             <h6>-1 : MD5-based password algorithm</h6>
             <h6>-crypt : standard Unix password algorithm</h6>
 
@@ -85,13 +88,18 @@
             <p>Use this tool to generate a new ssh/rsa keypair.</p>
             <p><a class="btn btn-secondary" href="/ssh" role="button">sshkeygen &raquo;</a></p>
           </div>
+          <div class="col-md-6">
+            <h2 class="text-primary">JSONPrettifier</h2>
+            <p>Use this tool to prettify/indent a complicated Json.</p>
+            <p><a class="btn btn-secondary" href="/jsonprettify" role="button">jsonprettify &raquo;</a></p>
+          </div>
         </div>
         <hr>
       </div> <!-- /container -->
     </main>
 
     <footer class="container">
-      <p>&copy; Tooling Portal 2019</p>
+      <p>&copy; Tooling Portal</p>
     </footer>
     <script src="../assets/custom/js/jquery-slim.min.js" ></script>
     <script>window.jQuery || document.write('<script src="../assets/custom/js/jquery-slim.min.js"><\/script>')</script>
