+++ web/templates/jsonprettify-process.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8">
     <link rel="icon" href="../assets/custom/img/tools.svg">
-    <title>HTPasswordGenerator</title>
+    <title>Json Prettifier</title>
     <!-- Bootstrap core CSS -->
     <link href="../assets/bootstrap4/css/bootstrap.min.css" rel="stylesheet">
     <style>
@@ -50,42 +50,38 @@
     <main role="main">
       <div class="jumbotron" style="padding: 1.5rem 1.5rem;">
         <div class="container">
-          <h1 class="display-3">HTPasswordGenerator</h1>
-          <p>Use this tool to create your htpasswd</p>
+          <h1 class="display-3">Json Prettifier</h1>
         </div>
       </div>
 
-      <div class="container">
-        <form style="max-width: 730px;" name="htpasswd" onsubmit="return validateUserPass()" action="/htpasswd-process" method="POST" enctype="multipart/form-data">
-          <div class="form-group">
-            <label for="user">Username:</label>
-            <input type="text" class="form-control" id="user" name="username" placeholder="Your username in lowercase">
-            <small class="form-text text-muted">Use lowercase characters only.</small>
+      <div class="container" style="max-width: 1440px;">
+        <div class="container" style="max-width: 1440px;">
+          <div class="row" >
+            <div class="col" style="padding: 0px 5px 20px 0px;">
+              <p class="text-success text-center">Initial Json:</p>
+              <form name="initialJson" >
+                <div class="form-group">
+                  <label for="initialJson"></label>
+                  <textarea class="form-control" rows="20" id="initialJson" name="initialJson" readonly>{{ .Text}}</textarea>
+                </div>
+              </form>
+              <button type="button" class="btn btn-success" onclick="copyToClipboard('initialJson')">Copy to clipboard</button>
+            </div>
+            <div class="col" style="padding: 0px 0px 20px 5px;">
+              <p class="text-success text-center">Processed Json:</p>
+              <form name="processedJson">
+                <div class="form-group">
+                  <label for="processedJson"></label>
+                  <textarea class="form-control" rows="20" id="processedJson" name="processedJson" readonly>{{ .Result}}</textarea>
+                </div>
+              </form>
+              <button type="button" class="btn btn-success" onclick="copyToClipboard('processedJson')">Copy to clipboard</button>
+            </div>
+            
           </div>
-          <div class="form-group">
-            <label for="pass">Password:</label>
-            <input type="password" name="password" class="form-control" id="pass" placeholder="Your password" >
-            <small class="form-text text-muted">Password must contain between 8 and 64 characters.</small>
-          </div>
-          <div class="custom-control custom-checkbox">
-            <input type="checkbox" onclick="revealPassword('pass')" class="custom-control-input" id="showpass">
-            <label class="custom-control-label" for="showpass">Show Password</label>
-            <p></p>
-          </div>
-          <div class="form-group">
-            <label for="algorithm">Choose an algorithm:</label>
-            <select style="max-width: 570px;" class="custom-select" id="algorithm" name="algorithm">
-              <option value="-apr1">MD5-based password algorithm, Apache variant</option>
-              <option value="-1">MD5-based password algorithm</option>
-              <option value="-crypt">standard Unix password algorithm</option>
-            </select>
-          </div>
-          <p></p>
-          <p></p>
-          <button type="submit" class="btn btn-primary">Generate htpassword</button>
-          <hr style="max-width: 740px;">
-        </form>
-        
+          <button type="button" class="btn btn-primary btn-block" onclick="location.href='/jsonprettify'">Prettify a new json</button>
+          <hr>
+        </div>
       </div> <!-- /container -->
     </main>
