+++ web/templates/jsonprettify.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8">
     <link rel="icon" href="../assets/custom/img/tools.svg">
-    <title>HTPasswordGenerator</title>
+    <title>Json Prettifier</title>
     <!-- Bootstrap core CSS -->
     <link href="../assets/bootstrap4/css/bootstrap.min.css" rel="stylesheet">
     <style>
@@ -50,42 +50,26 @@
     <main role="main">
       <div class="jumbotron" style="padding: 1.5rem 1.5rem;">
         <div class="container">
-          <h1 class="display-3">HTPasswordGenerator</h1>
-          <p>Use this tool to create your htpasswd</p>
+          <h1 class="display-3">Json Prettifier</h1>
+          <p>Use this tool to prettify/indent a json file</p>
         </div>
       </div>
 
-      <div class="container">
-        <form style="max-width: 730px;" name="htpasswd" onsubmit="return validateUserPass()" action="/htpasswd-process" method="POST" enctype="multipart/form-data">
-          <div class="form-group">
-            <label for="user">Username:</label>
-            <input type="text" class="form-control" id="user" name="username" placeholder="Your username in lowercase">
-            <small class="form-text text-muted">Use lowercase characters only.</small>
+      <div class="container" style="max-width: 1440px;">
+        <div class="container" style="max-width: 1440px;">
+          <div class="row" >
+            <div class="col" style="padding: 20px;">
+              <h5>Insert json contents below:</h5>
+              <form name="jsonprettify" style="max-width: 90%;" onsubmit="return validateJson()" action="/jsonprettify-process" method="POST" enctype="multipart/form-data">
+                <div class="form-group">
+                  <label for="text"></label>
+                  <textarea class="form-control" rows="20" id="text" name="text"></textarea>
+                </div>
+                <button type="submit" class="btn btn-primary">Prettify json</button>
+              </form>
+            </div>
           </div>
-          <div class="form-group">
-            <label for="pass">Password:</label>
-            <input type="password" name="password" class="form-control" id="pass" placeholder="Your password" >
-            <small class="form-text text-muted">Password must contain between 8 and 64 characters.</small>
-          </div>
-          <div class="custom-control custom-checkbox">
-            <input type="checkbox" onclick="revealPassword('pass')" class="custom-control-input" id="showpass">
-            <label class="custom-control-label" for="showpass">Show Password</label>
-            <p></p>
-          </div>
-          <div class="form-group">
-            <label for="algorithm">Choose an algorithm:</label>
-            <select style="max-width: 570px;" class="custom-select" id="algorithm" name="algorithm">
-              <option value="-apr1">MD5-based password algorithm, Apache variant</option>
-              <option value="-1">MD5-based password algorithm</option>
-              <option value="-crypt">standard Unix password algorithm</option>
-            </select>
-          </div>
-          <p></p>
-          <p></p>
-          <button type="submit" class="btn btn-primary">Generate htpassword</button>
-          <hr style="max-width: 740px;">
-        </form>
-        
+        </div>
       </div> <!-- /container -->
     </main>
