+++ web/templates/passgen-process.html
@@ -40,6 +40,9 @@
           <li class="nav-item">
             <a class="nav-link" href="/ssh">SSHKeyGenerator</a>
           </li>
+          <li class="nav-item">
+            <a class="nav-link" href="/jsonprettify">JsonPrettifier</a>
+          </li>
         </ul>
       </div>
     </nav>
