+++ web/templates/ssh.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8">
     <link rel="icon" href="../assets/custom/img/tools.svg">
-    <title>SSH Corner</title>
+    <title>SSHKeyGenerator</title>
     <!-- Bootstrap core CSS -->
     <link href="../assets/bootstrap4/css/bootstrap.min.css" rel="stylesheet">
     <style>
@@ -40,6 +40,9 @@
           <li class="nav-item">
             <a class="nav-link" href="/ssh">SSHKeyGenerator</a>
           </li>
+          <li class="nav-item">
+            <a class="nav-link" href="/jsonprettify">JsonPrettifier</a>
+          </li>
         </ul>
       </div>
     </nav>
@@ -48,7 +51,7 @@
       <div class="jumbotron" style="padding: 1.5rem 1.5rem;">
         <div class="container">
           <h1 class="display-3">SSHKeyGenerator</h1>
-          <p>Use this tool to generate a new SSH/rsa keypair</p></p>
+          <p>Use this tool to generate a new SSH/rsa keypair</p>
         </div>
       </div>
       <div class="container" >
@@ -71,11 +74,20 @@
               <label for="pass">Key password:</label>
               <input type="password" name="password" class="form-control" id="pass" placeholder="Insert password" >
               <small class="form-text text-muted">Password must contain between 8 and 64 characters.</small>
-              <div class="custom-control custom-checkbox" id="pass">
+              <div class="custom-control custom-checkbox" id="showpassword">
                 <input type="checkbox" onclick="revealPassword('pass')" class="custom-control-input" id="showpass">
                 <label class="custom-control-label" for="showpass">Show Password</label>
                 <p></p>
               </div>
+            </div>
+            <div class="form-group">
+              <label for="bitsize">Select the key bitsize:</label>
+              <select style="max-width: 570px;" class="custom-select" id="bitsize" name="bitsize">
+                <option value="4096">4096</option>
+                <option value="3072">3072</option>
+                <option value="1024">1024</option>
+              </select>
+              <small class="form-text text-muted">Specifies the number of bits in the key to create.</small>
             </div>
             <p></p>
             <p></p>
